requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.2
django-admin-black==1.0.9
django-browser-reload==1.12.1
djangorestframework==3.14.0
exceptiongroup==1.2.1
fernet==1.0.1
h11==0.14.0
html2text==2024.2.26
idna==3.6
install==1.3.5
mysqlclient==2.2.1
outcome==1.3.0.post0
packaging==24.1
pyaes==1.6.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
str2bool==1.1
trio==0.25.1
trio-websocket==0.11.1
typing-extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
